Move form processor case models to module
